updateprize1

+++ 代码/iMidudu.Lucky/iMidudu.Lucky.Web/WebService.asmx.cs
@@ -71,7 +71,7 @@ namespace iMidudu.Lucky.Web
                 iMidudu.Lucky.Web.SystemDAO.SqlHelper.ExecteNonQueryProcedure("PrizeNameUpdate_Procedure",
                      new System.Data.SqlClient.SqlParameter("@PrizeId", d.PrizeId),
                      new System.Data.SqlClient.SqlParameter("@PrizeName", d.PrizeName),
-                     new System.Data.SqlClient.SqlParameter("@Quantity", (int)d.Quantity)
+                     new System.Data.SqlClient.SqlParameter("@Quantity", d.Quantity)
                      );
 
             }
